Move the session-not-found test into session_store_test.go
